+++ src/app/app.module.ts
@@ -12,8 +12,8 @@ import { ButtonsModule } from 'ngx-bootstrap/buttons';
 import { CollapseModule } from 'ngx-bootstrap/collapse';
 import { TooltipModule } from 'ngx-bootstrap/tooltip';
 import { MoneyAddCalcComponent } from './money/calculator/money-add-calc.component';
-import { MoneyCategoryPickerComponent } from './money/category-picker/money-cat-picker.component';
-import { MoneyCategoryFilterComponent } from './money/category-picker/money-cat-filter.component';
+import { MoneyCategoryPickerComponent } from './money/category/money-cat-picker.component';
+import { MoneyCategoryFilterComponent } from './money/category/money-cat-filter.component';
 import { MoneyListComponent } from './money/money-list.component';
 import { HouseComponent } from './house/house.component';
 import { WeightComponent } from './weight/weight.component';
