+++ src/app/money/list-row-line/list-row-line-factory.ts
@@ -7,12 +7,12 @@ import {ListRowLineTotalCredits} from './list-row-line-total-credits';
 import {ListRowLineTotalDebits} from './list-row-line-total-debits';
 import {ListRowLineRegular} from './list-row-line-regular';
 import {ListRowLineReconcileTop} from './list-row-line-reconcile-top';
-import {IMatch} from '../money-match';
+import {IMatch} from '../reconciliation/match';
 import {IRegular} from '../money-regular';
 import {MoneyService} from '../money.service';
 import {IStatement} from '../money-statement';
 import {ListRowSummary} from './list-row-summary';
-import {ITransaction} from '../money-transaction'
+import {ITransaction} from '../transaction/transaction'
 import {ListRowLineHeader} from "./list-row-line-header";
