+++ src/app/money/list-row-line/list-row-line-total-cfwd.ts
@@ -1,5 +1,5 @@
 import {IListRowLineInterface, ListRowLineType} from './list-row-line-interface';
-import {ICategory} from '../money-category';
+import {ICategory} from '../category/money-category';
 import {IStatement} from '../money-statement';
 import {MoneyService} from '../money.service';
 import {ListRowSummary} from './list-row-summary';
